requirements.txt
streamlit
snowflake-connector-python
python-dotenv
pandas
plotly